google-adk/beginner/ai_resume_evaluator_agent/requirements.txt
google-adk 